Avoid failing if no .gitmodules file present

+++ tools/deploy.sh
@@ -84,7 +84,9 @@ setup_gh() {
 backup() {
   mv "$SITE_DIR"/* "$_backup_dir"
   mv .git "$_backup_dir"
-  mv .gitmodules "$_backup_dir"
+  if [[ -f ".gitmodules" ]]; then
+    mv .gitmodules "$_backup_dir"
+  fi
 
   # When adding custom domain from Github website,
   # the CANME only exist on `gh-pages` branch
